add SensorPage temperature block

+++ app/src/main/java/se/ifmo/ru/smartapp/ui/data/Sensors.kt
@@ -4,7 +4,7 @@ data class Sensor(
     val id: Long,
     val name: String,
     val type: String,
-    var value: Long,
+    var value: Double,
     var stateId: Long,
     val roomId: Long
 )
